requirements.txt
torch==1.7.1
torchvision==0.8.2
Pillow==8.1.0
pytest==6.2.2